misc: move gateway and relay to networking page
