+++ frontend/src/pages/organization/NetworkingPage/components/GatewayTab/index.tsx
@@ -1,0 +1,1 @@
+export { GatewayTab } from "./GatewayTab";
