+++ frontend/src/pages/organization/NetworkingPage/components/NetworkingTabGroup/index.tsx
@@ -1,0 +1,1 @@
+export { NetworkingTabGroup } from "./NetworkingTabGroup";
