+++ frontend/src/pages/organization/NetworkingPage/components/RelayTab/index.tsx
@@ -1,0 +1,1 @@
+export { RelayTab } from "./RelayTab";
