+++ frontend/src/routes.ts
@@ -40,8 +40,7 @@ const organizationRoutes = route("/organization", [
       "organization/AppConnections/OauthCallbackPage/route.tsx"
     )
   ]),
-  route("/gateways", [index("organization/Gateways/GatewayListPage/route.tsx")]),
-  route("/relays", [index("organization/Gateways/Relay/RelayListPage/route.tsx")])
+  route("/networking", [index("organization/NetworkingPage/route.tsx")])
 ]);
 
 const secretManagerRoutes = route("/projects/secret-management/$projectId", [
